requirements.txt
# Core dependencies
langchain>=0.0.267
requests>=2.28.0

# Testing
pytest>=7.0.0
python-dotenv>=1.0.0

# Development
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
build>=0.10.0
twine>=4.0.0

# Example dependencies
langgraph>=0.0.15
langgraph-prebuilt>=0.0.5